convert single-tag panel to new explorer API
